+++ src/Form/TopicType.php
@@ -1,0 +1,46 @@
+<?php
+
+namespace App\Form;
+
+use App\Entity\Topic;
+use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\Extension\Core\Type\FileType;
+use Symfony\Component\Form\Extension\Core\Type\SubmitType;
+use Symfony\Component\Form\Extension\Core\Type\TextareaType;
+use Symfony\Component\Form\Extension\Core\Type\TextType;
+use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\OptionsResolver\OptionsResolver;
+
+class TopicType extends AbstractType
+{
+    public function buildForm(FormBuilderInterface $builder, array $options): void
+    {
+        $builder
+            ->add('title', TextType::class, [
+                'attr' => [
+                    'placeholder' => 'Titre du topic'
+                ]
+            ])
+            ->add('content', TextareaType::class, [
+                'attr' => [
+                    'placeholder' => 'Contenu du topic'
+                ]
+            ])
+            ->add('image', FileType::class, [
+                'label' => 'Image du topic',
+                'mapped' => false,
+                'required' => false
+            ])
+            ->add('submit', SubmitType::class, [
+                'label' => 'Créer un topic'
+            ])
+        ;
+    }
+
+    public function configureOptions(OptionsResolver $resolver): void
+    {
+        $resolver->setDefaults([
+            'data_class' => Topic::class,
+        ]);
+    }
+}